Moved exclusive mode related files into separated folder
