+++ src/ExclusiveMode/InputExclusiveModeImpl.h
