+++ src/ExclusiveMode/LinuxExclusiveMode.h
