+++ src/ExclusiveMode/MacOSExclusiveMode.cpp
