+++ src/ExclusiveMode/MacOSExclusiveMode.h
