+++ src/ExclusiveMode/WindowsExclusiveMode.cpp
