+++ src/ExclusiveMode/WindowsExclusiveMode.h
